+++ src/cli_parser.rs
@@ -23,7 +23,13 @@ pub const FLAG_CONFIG: &str = "config";
 
 pub fn parse_arguments<'a>() -> clap::ArgMatches<'a> {
     let app = App::new("cinit")
-        .version(env!("CARGO_PKG_VERSION"))
+        .version(concat!(
+            env!("CARGO_PKG_VERSION"),
+            " ",
+            env!("VERGEN_SHA"),
+            " ",
+            env!("VERGEN_BUILD_TIMESTAMP")
+        ))
         .about(env!("CARGO_PKG_DESCRIPTION"))
         .arg(
             Arg::with_name("config")