+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+* `cinit --version` also prints the commit hash and date of the build
+
 ## [1.3.8]
 
 ### Maintenance